Develop

+++ lib/ui/settings/settings_page.dart
@@ -16,8 +16,10 @@ import 'package:vocechat_client/services/chat_service.dart';
 import 'package:vocechat_client/services/db.dart';
 import 'package:vocechat_client/ui/app_colors.dart';
 import 'package:vocechat_client/ui/auth/server_page.dart';
+import 'package:vocechat_client/ui/chats/chats/chats_main_page.dart';
 import 'package:vocechat_client/ui/settings/firebase_settings_page.dart';
 import 'package:vocechat_client/ui/settings/server_info_settings_page.dart';
+import 'package:vocechat_client/ui/settings/settings_about_page.dart';
 import 'package:vocechat_client/ui/settings/settings_bar.dart';
 import 'package:vocechat_client/ui/settings/userinfo_setting_page.dart';
 import 'package:vocechat_client/ui/widgets/app_banner_button.dart';
@@ -75,10 +77,11 @@ class _SettingPageState extends State<SettingPage> {
                 children: [
                   _buildUserInfo(),
                   _buildServer(context),
-                  if (App.app.userDb?.userInfo.isAdmin ?? false)
-                    _buildConfigs(context),
                   _buildAbout(),
-                  SizedBox(height: 8),
+                  if (App.app.userDb?.userInfo.isAdmin ?? false)
+                    // _buildConfigs(context),
+
+                    SizedBox(height: 8),
                   _buildButtons(context)
                 ],
               )),
@@ -143,11 +146,29 @@ class _SettingPageState extends State<SettingPage> {
   }
 
   Widget _buildAbout() {
+    // return BannerTile(
+    //     title: AppLocalizations.of(context)!.settingsPageAbout,
+    //     keepArrow: false,
+    //     enableTap: false,
+    //     trailing: FutureBuilder<String>(
+    //         future: _getVersion(),
+    //         builder: (context, snapshot) {
+    //           if (snapshot.hasData) {
+    //             return Text(snapshot.data!,
+    //                 style: TextStyle(
+    //                     fontSize: 15,
+    //                     fontWeight: FontWeight.w400,
+    //                     color: AppColors.grey500));
+    //           } else {
+    //             return SizedBox.shrink();
+    //           }
+    //         }));
     return BannerTile(
-        header: AppLocalizations.of(context)!.settingsPageAbout,
-        enableTap: false,
-        title: AppLocalizations.of(context)!.settingsPageAboutVersion,
-        keepArrow: false,
+        title: AppLocalizations.of(context)!.settingsPageAbout,
+        keepArrow: true,
+        enableTap: true,
+        onTap: () => Navigator.of(context)
+            .push(MaterialPageRoute(builder: (context) => SettingsAboutPage())),
         trailing: FutureBuilder<String>(
             future: _getVersion(),
             builder: (context, snapshot) {
@@ -195,10 +216,11 @@ class _SettingPageState extends State<SettingPage> {
             },
             title: AppLocalizations.of(context)!.settingsPageClearData),
         SizedBox(height: 8),
-        AppBannerButton(
-          onTap: () => _onDeleteAccountTapped(context),
-          title: "Delete Account",
-        )
+        if (App.app.userDb?.uid != 1)
+          AppBannerButton(
+            onTap: () => _onDeleteAccountTapped(context),
+            title: "Delete Account",
+          )
       ],
     );
   }
@@ -261,13 +283,10 @@ class _SettingPageState extends State<SettingPage> {
               final api = UserApi(App.app.chatServerM.fullUrl);
               final res = await api.delete();
               if (res.statusCode == 200) {
-                App.app.authService!.selfDelete().then((value) {
+                App.app.authService!.selfDelete().then((value) async {
                   try {
-                    navigatorKey.currentState!.pushAndRemoveUntil(
-                        MaterialPageRoute(
-                          builder: (context) => ServerPage(),
-                        ),
-                        (route) => false);
+                    await App.app.changeUserAfterLogOut();
+                    navigatorKey.currentState!.pop();
                   } catch (e) {
                     App.logger.severe(e);
                   }
@@ -318,7 +337,8 @@ class _SettingPageState extends State<SettingPage> {
     }
   }
 
-  Future<void> _onUser(UserInfoM userInfoM, EventActions action) async {
+  Future<void> _onUser(
+      UserInfoM userInfoM, EventActions action, bool afterReady) async {
     if (userInfoM.uid == App.app.userDb?.uid) {
       userInfoNotifier.value = userInfoM;
